% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eps.R
\name{get_eps}
\alias{get_eps}
\title{Get the errors with specified distribution.}
\usage{
get_eps(eps_dist, n)
}
\arguments{
\item{eps_dist}{specified type of distibution.Supports ("uniform","t","laplace").}

\item{n}{sample size}
}
\value{
numeric vector.
}
\description{
\code{get_eps} returns a randomly generated numeric vector.
}
\examples{
get_eps(eps_dist = "uniform", n = 100)
}
