% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_url.R
\name{make_url}
\alias{make_url}
\title{Form url for API call to recipe puppy}
\usage{
make_url(user_ingredients_input, user_search_query)
}
\arguments{
\item{user_ingredients_input}{vector of character strings.}

\item{user_search_query}{a character string.}
}
\value{
ulr as a character string.
}
\description{
\code{make_url} returns a url as a character string based in user inputs. Made to handle NULL.
}
\examples{
make_url(c("onion", "cheese", "beef"), "burger")
}
