% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regsim.R
\name{plot_regsim}
\alias{plot_regsim}
\title{Plots the Simulated Regression Results}
\usage{
plot_regsim(reps = 1000, beta = seq(-1, 1, by = 0.1), n = c(10, 20, 30,
  40, 50), eps_dist = "uniform")
}
\arguments{
\item{reps}{nuber of simulated regressions. Default value of 1000.}

\item{beta}{vector of coeficients. Default sequence from -1 to 1  by = 0.1.}

\item{n}{vector of sample sizes. Default values set to (10, 20, 30, 40, 50).}

\item{eps_dist}{destribution of error term. Supports ("uniform","t","laplace"). Default set to "uniform".}
}
\value{
plots.
}
\description{
\code{plot_regsim} returns a list with simulated results.
}
\examples{
plot_regsim(reps = 1000, n = c(10, 20, 30, 40, 50), beta = seq(-1, 1, by = 0.1), eps_dist = "uniform")

}
