% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_sim.R
\name{reg_sim}
\alias{reg_sim}
\title{Simulates Regression with non-Normal and Normal error terms.}
\usage{
reg_sim(reps = 1000, n = 100, beta = 0.5, eps_dist = "uniform")
}
\arguments{
\item{reps}{number of simulated regressions. Default value of 1000.}

\item{n}{sample size. Default value of 100.}

\item{beta}{the slope coeficient. Default value of 0.5.}

\item{eps_dist}{distribution of error term. Supports ("uniform","t","laplace"). Default set to "uniform".}
}
\value{
list of estimated beta coefficients, R^2, and power estimation for each repetition, for both
errors of a specified distribution and normal errors.
}
\description{
\code{reg_sim} returns a list with simulated results.
}
\examples{
reg_sim(reps = 1000, n = 100, beta = 0.1, eps_dist = "uniform")

}
