% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim.R
\name{regsim}
\alias{regsim}
\title{Simulates Regression with non-Normal error terms.}
\usage{
regsim(reps = 1000, n = 100, beta = 0.5, eps_dist = "uniform")
}
\arguments{
\item{reps}{nuber of simulated regressions. Default value of 1000.}

\item{n}{sample size. Default value of 100.}

\item{beta}{the slope coeficient. Default value of 0.5.}

\item{eps_dist}{distribution of error term. Supports ("uniform","t","laplace"). Default set to "uniform".}
}
\value{
list of mean estimated beta coefficient, mean estimated R^2, and estimated power for each repetition.
Returns results for the specified distribution only.
}
\description{
\code{regsim} returns a list with simulated results.
}
\examples{
regsim(reps = 1000, n = 100, beta = 0.1, eps_dist = "uniform")
}
