% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_recipe.R
\name{search_recipe}
\alias{search_recipe}
\title{Searching Recipes}
\usage{
search_recipe(ingredients = NULL, search = NULL, minNumIngredients = 1)
}
\arguments{
\item{ingredients}{vector of character strings.}

\item{search}{a character string.}

\item{minNumIngredients}{an integer.}
}
\value{
data.frame object with search results
}
\description{
\code{search_recipe} returns a data.frame with search results. Make sure to spell the ingredients correctly,
or the search won't work. You must specify at leat one ingredient. You can also add a search for a particular type
of dish via the `search` parameter. If you pass on more than 2 ingredients, you may set a minimum number of
supplied ingredients to be included in the recipe. Ingredients more important for the dish must be passed in first.
The function returns a data.frame object with a list of links to recipes, required ingredients, and name of the dish.
Bon Appetit!
}
\examples{
search_recipe(c("onion", "cheese", "beef"), "burger", 3)

}
